Disables telemetry collection
Disables telemetry collection by commenting out the PostHog client initialization and setting telemetryEnabled to false.
This prevents data from being sent to PostHog.
Opt-in and opt-out are still called, but on an uninitialized client.

Removes dotenvx and telemetryClient as they are not currently needed.

telemetryClient needs further investigation to determine if it is used instead of telemetryService.